fix: refatoring Detalhes Comida
